+++ commands/dumpsmc/dumpsmc.go
@@ -10,6 +10,16 @@ import (
 	"github.com/drdonk/golocker/vmwpatch"
 )
 
+func waitExit() {
+	quiet := os.Getenv("UNLOCK_QUIET")
+	if quiet != "1" {
+		fmt.Println()
+		fmt.Println("\nPress Enter key to continue...")
+
+		_, _ = fmt.Scanln()
+	}
+}
+
 func main() {
 	vmwpatch.IsAdmin()
 
@@ -21,6 +31,7 @@ func main() {
 	var filename string
 	if len(os.Args) < 2 {
 		fmt.Printf("Please pass a file name!\n")
+		waitExit()
 		return
 	} else {
 		filename = os.Args[1]
@@ -29,5 +40,6 @@ func main() {
 	vmwpatch.DumpSMC(filename)
 	patchFlag, patchStatus, hash256 := vmwpatch.IsSMCPatched(filename)
 	fmt.Printf("\nPatch Status: %s (%d)\nSHA256: %s\n", patchStatus, patchFlag, hash256)
+	waitExit()
 
 }