+++ debug-build.sh
@@ -1,24 +1,77 @@
-#!/usr/bin/env bash
+#!/usr/bin/env zsh
+#set -x
+
 echo Building debug executables
+if ! [ $# -eq 1 ] ; then
+  echo "Product version not found" >&2
+  exit 1
+fi
 
-echo Building Windows executables...
-env GOOS=windows GOARCH=amd64 go build -o ./windows/dumpsmc.exe ./tools/dumpsmc.go
-env GOOS=windows GOARCH=amd64 go build -o ./windows/patchsmc.exe ./tools/patchsmc.go
-env GOOS=windows GOARCH=amd64 go build -o ./windows/patchgos.exe ./tools/patchgos.go
-env GOOS=windows GOARCH=amd64 go build -o ./windows/patchvmkctl.exe ./tools/patchvmkctl.go
-env GOOS=windows GOARCH=amd64 go build -o ./windows/unlocker.exe ./command/unlocker.go
+mkdir -p ./dist/macos
 
-echo Building Linux executables...
-env GOOS=linux GOARCH=amd64 go build -o ./linux/dumpsmc ./tools/dumpsmc.go
-env GOOS=linux GOARCH=amd64 go build -o ./linux/patchsmc ./tools/patchsmc.go
-env GOOS=linux GOARCH=amd64 go build -o ./linux/patchgos ./tools/patchgos.go
-env GOOS=linux GOARCH=amd64 go build -o ./linux/patchvmkctl ./tools/patchvmkctl.go
-env GOOS=linux GOARCH=amd64 go build -o ./linux/unlocker ./command/unlocker.go
+pushd ./commands/check
+echo "Building check"
+go-winres make --arch amd64 --product-version $1 --file-version $1
+env GOOS=windows GOARCH=amd64 go build -o ../../dist/windows
+env GOOS=linux GOARCH=amd64 go build -o ../../dist/linux
+rm rsrc_windows_amd64.syso
+popd
 
-echo Building macOS executables...
-env GOOS=darwin GOARCH=amd64 go build -o ./darwin/dumpsmc ./tools/dumpsmc.go
-env GOOS=darwin GOARCH=amd64 go build -o ./darwin/patchsmc ./tools/patchsmc.go
-env GOOS=darwin GOARCH=amd64 go build -o ./darwin/patchgos ./tools/patchgos.go
-env GOOS=darwin GOARCH=amd64 go build -o ./darwin/patchvmkctl ./tools/patchvmkctl.go
-env GOOS=darwin GOARCH=amd64 go build -o ./darwin/unlocker ./command/unlocker.go
+pushd ./commands/relock
+echo "Building relock"
+go-winres make --arch amd64 --product-version $1 --file-version $1
+env GOOS=windows GOARCH=amd64 go build -o ../../dist/windows
+env GOOS=linux GOARCH=amd64 go build -o ../../dist/linux
+env GOOS=darwin GOARCH=amd64 go build -o ../../dist/macos
+rm rsrc_windows_amd64.syso
+popd
 
+pushd ./commands/unlock
+echo "Building unlock"
+go-winres make --arch amd64 --product-version $1 --file-version $1
+env GOOS=windows GOARCH=amd64 go build -o ../../dist/windows
+env GOOS=linux GOARCH=amd64 go build -o ../../dist/linux
+env GOOS=darwin GOARCH=amd64 go build -o ../../dist/macos
+rm rsrc_windows_amd64.syso
+popd
+
+pushd ./commands/dumpsmc
+echo "Building dumpsmc"
+go-winres make --arch amd64 --product-version $1 --file-version $1
+env GOOS=windows GOARCH=amd64 go build -o ../../dist/windows
+env GOOS=linux GOARCH=amd64 go build -o ../../dist/linux
+env GOOS=darwin GOARCH=amd64 go build -o ../../dist/macos
+rm rsrc_windows_amd64.syso
+popd
+
+pushd ./commands/patchgos
+echo "Building patchgos"
+go-winres make --arch amd64 --product-version $1 --file-version $1
+env GOOS=windows GOARCH=amd64 go build -o ../../dist/windows
+env GOOS=linux GOARCH=amd64 go build -o ../../dist/linux
+env GOOS=darwin GOARCH=amd64 go build -o ../../dist/macos
+rm rsrc_windows_amd64.syso
+popd
+
+pushd ./commands/patchsmc
+echo "Building patchsmc"
+go-winres make --arch amd64 --product-version $1 --file-version $1
+env GOOS=windows GOARCH=amd64 go build -o ../../dist/windows
+env GOOS=linux GOARCH=amd64 go build -o ../../dist/linux
+env GOOS=darwin GOARCH=amd64 go build -o ../../dist/macos
+rm rsrc_windows_amd64.syso
+popd
+
+pushd ./commands/patchvmkctl
+echo "Building patchvmkctl"
+go-winres make --arch amd64 --product-version $1 --file-version $1
+env GOOS=windows GOARCH=amd64 go build -o ../../dist/windows
+env GOOS=linux GOARCH=amd64 go build -o ../../dist/linux
+env GOOS=darwin GOARCH=amd64 go build -o ../../dist/macos
+rm rsrc_windows_amd64.syso
+popd
+
+cp -v LICENSE ./dist
+cp -v *.md ./dist
+cp -v *.pdf ./dist
+cp -vr ./iso ./dist