+++ vmwpatch/vmw_windows.go
@@ -12,8 +12,6 @@ import (
 	"os/exec"
 	"path/filepath"
 	"runtime"
-	"strings"
-	"syscall"
 	"time"
 
 	"golang.org/x/sys/windows"
@@ -25,22 +23,7 @@ import (
 var manager *mgr.Mgr
 
 func IsAdmin() bool {
-	_, err := os.Open("\\\\.\\PHYSICALDRIVE0")
-	if err != nil {
-		exe, _ := os.Executable()
-		cwd, _ := os.Getwd()
-
-		verbPtr, _ := syscall.UTF16PtrFromString("runas")
-		exePtr, _ := syscall.UTF16PtrFromString(exe)
-		cwdPtr, _ := syscall.UTF16PtrFromString(cwd)
-		argPtr, _ := syscall.UTF16PtrFromString(strings.Join(os.Args[1:], " "))
-
-		err := windows.ShellExecute(0, verbPtr, exePtr, argPtr, cwdPtr, 1)
-		if err != nil {
-			fmt.Println(err)
-		}
-		os.Exit(1)
-	}
+	// Always return true as now has embedded manifest to runas Administrator
 	return true
 }
 
@@ -55,7 +38,11 @@ func VMWStart(v *VMwareInfo) {
 		fmt.Println("Disconnect from SCM failed")
 		// Not stopping the process over this one
 	}
+
+	cwd, _ := os.Getwd()
+	os.Chdir("C:\\Windows\\SysWOW64\\")
 	taskStart(filepath.Join(v.InstallDir, v.Tray))
+	os.Chdir(cwd)
 }
 
 func VMWStop(v *VMwareInfo) {
@@ -92,27 +79,32 @@ func VMWInfo() *VMwareInfo {
 	regKey, err := registry.OpenKey(registry.LOCAL_MACHINE,
 		`SOFTWARE\VMware, Inc.\VMware Player`, access)
 	if err != nil {
-		panic("Failed to open registry")
+		fmt.Println("Failed to open VMware registry key")
+		return v
 	}
 	//goland:noinspection GoUnhandledErrorResult
 	defer regKey.Close()
 
 	v.ProductVersion, _, err = regKey.GetStringValue("ProductVersion")
 	if err != nil {
-		panic("Failed to locate registry key ProductVersion")
+		fmt.Println("Failed to locate registry key ProductVersion")
+		return v
 	}
 
 	v.BuildNumber, _, err = regKey.GetStringValue("BuildNumber")
 	if err != nil {
-		panic("Failed to locate registry key BuildNumber")
+		fmt.Println("Failed to locate registry key BuildNumber")
+		return v
 	}
 
 	v.InstallDir, _, err = regKey.GetStringValue("InstallPath")
 	if err != nil {
-		panic("Failed to locate registry key InstallPath")
+		fmt.Println("Failed to locate registry key InstallPath")
+		return v
 	}
 
 	// Construct needed filenames from reg settings
+	v.BasePath = getBaseDir()
 	v.InstallDir64 = filepath.Join(v.InstallDir, "x64")
 	v.Player = "vmplayer.exe"
 	v.Workstation = "vmware.exe"
@@ -128,20 +120,21 @@ func VMWInfo() *VMwareInfo {
 	v.PathVMXDebug = filepath.Join(v.InstallDir64, v.VMXDebug)
 	v.PathVMXStats = filepath.Join(v.InstallDir64, v.VMXStats)
 	v.PathVMwareBase = filepath.Join(v.InstallDir, v.VMwareBase)
-	currentFolder, _ := os.Getwd()
-	v.BackDir = filepath.Join(currentFolder, "backup", v.ProductVersion)
+	v.BackDir = filepath.Join(v.BasePath, "backup", v.ProductVersion)
 	v.BackVMXDefault = filepath.Join(v.BackDir, v.VMXDefault)
 	v.BackVMXDebug = filepath.Join(v.BackDir, v.VMXDebug)
 	v.BackVMXStats = filepath.Join(v.BackDir, v.VMXStats)
 	v.BackVMwareBase = filepath.Join(v.BackDir, v.VMwareBase)
-	v.PathISOMacOSX = filepath.Join(v.InstallDir, "darwinPre15.iso")
-	v.PathISOmacOS = filepath.Join(v.InstallDir, "darwin.iso")
+	v.SrcISOMacOSX = filepath.Join(v.BasePath, "iso", "darwinPre15.iso")
+	v.SrcISOmacOS = filepath.Join(v.BasePath, "iso", "darwin.iso")
+	v.DstISOMacOSX = filepath.Join(v.InstallDir, "darwinPre15.iso")
+	v.DstISOmacOS = filepath.Join(v.InstallDir, "darwin.iso")
 	return v
 }
 
 func setCTime(path string, ctime time.Time) error {
-	//setCTime will set the create time on a file. On Windows, this requires
-	//calling SetFileTime and explicitly including the create time.
+	//setCTime will set the creation time on a file. On Windows, this requires
+	//calling SetFileTime and explicitly including the creation time.
 	ctimespec := windows.NsecToTimespec(ctime.UnixNano())
 	pathp, e := windows.UTF16PtrFromString(path)
 	if e != nil {