+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All dates are UK format.
 
+## 13/05/22 4.2.0
+_drdonk:_
+* Split unlocker.exe into 2 new files unlock.exe and relock.exe so can be directly executed on Windows without 
+having to pass a parameter to install or uninstall
+* Added Windows manifest to allow the executables to directly run as Administrator without using re-spawn code 
+* Minimize false AV detection especially Windows Defender
+* Added a new command to check the status of the patching called check.exe
+* Fix to restarting vmware-tray.exe on Windows which had the wrong working directory
+
 ## 03/03/22 4.1.3
 _drdonk:_
 * Update VMware Tools to version 12.0.0